chore(across): extract logic into common
