+++ src/apps/across/ethereum/across.pool.token-fetcher.ts
@@ -1,24 +1,8 @@
-import { Inject } from '@nestjs/common';
-
-import { APP_TOOLKIT, IAppToolkit } from '~app-toolkit/app-toolkit.interface';
 import { Register } from '~app-toolkit/decorators';
-import { AppTokenTemplatePositionFetcher } from '~position/template/app-token.template.position-fetcher';
-import {
-  GetAddressesParams,
-  GetDataPropsParams,
-  GetPriceParams,
-  GetPricePerShareParams,
-  GetUnderlyingTokensParams,
-} from '~position/template/app-token.template.types';
 import { Network } from '~types/network.interface';
 
 import ACROSS_DEFINITION from '../across.definition';
-import { AcrossContractFactory, BadgerPool } from '../contracts';
-
-export type AcrossPoolTokenDefinition = {
-  address: string;
-  underlyingTokenAddress: string;
-};
+import { AcrossPoolTokenDefinition, AcrossPoolTokenFetcher } from '../common/across.pool.token-fetcher';
 
 const v1Pools = [
   {
@@ -86,50 +70,13 @@ const groupId = ACROSS_DEFINITION.groups.pool.id;
 const network = Network.ETHEREUM_MAINNET;
 
 @Register.TokenPositionFetcher({ appId, groupId, network })
-export class EthereumAcrossPoolTokenFetcher extends AppTokenTemplatePositionFetcher<BadgerPool> {
+export class EthereumAcrossPoolTokenFetcher extends AcrossPoolTokenFetcher {
   appId = appId;
   groupId = groupId;
   network = network;
   groupLabel = 'Pools';
 
-  constructor(
-    @Inject(APP_TOOLKIT) protected readonly appToolkit: IAppToolkit,
-    @Inject(AcrossContractFactory) protected readonly contractFactory: AcrossContractFactory,
-  ) {
-    super(appToolkit);
-  }
-
-  getContract(address: string): BadgerPool {
-    return this.contractFactory.badgerPool({ network: this.network, address });
-  }
-
   async getDefinitions(): Promise<AcrossPoolTokenDefinition[]> {
     return [...v1Pools, ...v2Pools];
-  }
-
-  async getAddresses({ definitions }: GetAddressesParams<AcrossPoolTokenDefinition>): Promise<string[]> {
-    return definitions.map(v => v.address);
-  }
-
-  async getUnderlyingTokenAddresses({ definition }: GetUnderlyingTokensParams<BadgerPool, AcrossPoolTokenDefinition>) {
-    return definition.underlyingTokenAddress;
-  }
-
-  async getPricePerShare({ contract, appToken, multicall }: GetPricePerShareParams<BadgerPool>) {
-    const pricePerShareRaw = await multicall.wrap(contract).callStatic.exchangeRateCurrent();
-    const decimals = appToken.tokens[0].decimals;
-
-    return Number(pricePerShareRaw) / 10 ** decimals;
-  }
-
-  async getPrice({ appToken }: GetPriceParams<BadgerPool>) {
-    return appToken.tokens[0].price * Number(appToken.pricePerShare);
-  }
-
-  async getDataProps(opts: GetDataPropsParams<BadgerPool>) {
-    const { appToken } = opts;
-    const liquidity = appToken.price * appToken.supply;
-
-    return { liquidity };
   }
 }
